lots of styling, improving names
