+++ components/MainContentContainer.vue
@@ -1,8 +1,11 @@
+
+import SearchResultDetails from './SearchResultDetails.vue';
+
 <template>
     <div class="flex flex-1 bg-secondary-bg/20 hover:shadow-inner hover:shadow-secondary-bg/90">
         <!-- <WelcomeSection /> -->
         <Modal class="h-1/2 ml-8 mt-12">
-            <ResultDetails />
+            <SearchResultDetails />
         </Modal>
         <MapContainer />
     </div>
